requirements.txt
playwright==1.42.0
pandas==2.2.1
openpyxl==3.1.2
supabase==2.3.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
python-dotenv==1.0.1
requests==2.31.0
aiohttp==3.9.3
asyncio==3.4.3
colorama==0.4.6 